zacks-cdk-library/requirements-dev.txt
pytest~=7.0.0
pytest-cov~=4.0.0
black~=23.0.0
isort~=5.12.0
flake8~=6.0.0
mypy~=1.0.0
aws-cdk-lib>=2.0.0
constructs>=10.0.0